Conflicts:
	SecureWitness/Report/forms.py
	SecureWitness/Report/models.py
	SecureWitness/Report/views.py
	SecureWitness/accounts/models.py
	SecureWitness/accounts/views.py
	SecureWitness/myDataBase.db
	SecureWitness/templates/add_folder.html
	SecureWitness/templates/index.html
	SecureWitness/templates/report_list.html
	SecureWitness/templates/user_profile.html

+++ SecureWitness/accounts/views.py
@@ -10,7 +10,9 @@ from django.db.models import Q
 from accounts.models import UserProfile, UserGroup
 from Report.models import Folder
 from Report.models import reports
-from accounts.forms import GroupCreationForm
+from accounts.forms import GroupCreationForm, UserGroupCreationForm
+from django.contrib.auth.decorators import login_required
+
 
 # Create your views here.
 
@@ -43,15 +45,59 @@ def register(request, creation_form=UserCreationForm,extra_context=None):
 
 def profile(request):
     folders = Folder.objects.all()[:20]
-    return render(request, 'user_profile.html', {'folder': folders})
+    if len(folders) == 0:
+        folder = Folder()
+        folder.save()
+        # folders = Folder.objects.all()[:20]
+    # This code allows for an admin link on the user profile page.
+    if request.user.is_active:
+        profile = UserProfile.objects.filter(user = request.user)[0]
+    else:
+        profile = None
+
+    root_folder = Folder.objects.get(pk=1)
+    return render(request, 'user_profile.html', {'o': root_folder, 'prof': profile})
 
 
 def report_list(request, folder_id):
     return render(request, 'report_list.html', {'folder': Folder.objects.all().filter(pk=folder_id)[0]})
 
+
+def add_folder(request):
+    if request.method == 'POST':
+        title = request.POST.get("file_name")
+
+        # handel invalid input
+        if len(title) == 0:
+            error = "Error: folder name needed!"
+            folders = Folder.objects.all()[:20]
+            return render(request, 'add_folder.html', {'folder': folders, 'message': error})
+        elif Folder.objects.filter(file_name=title).count() != 0:
+            error = "Error: folder already exist!"
+            folders = Folder.objects.all()[:20]
+            return render(request, 'add_folder.html', {'folder': folders, 'message': error})
+        else:
+            parent_name = request.POST.get("parent_folder")
+            parent = Folder.objects.get(file_name=parent_name)
+            folder = Folder(file_name=title, parent_folder=parent)
+            folder.save()
+
+        entries = Folder.objects.all()[:20]
+
+        if request.user.is_active:
+            profile = UserProfile.objects.filter(user = request.user)[0]
+        else:
+            profile = None
+
+        return render(request, 'user_profile.html', {'folder': entries, 'prof': profile})
+
+    folders = Folder.objects.all()[:20]
+    error = None
+    return render(request, 'add_folder.html', {'folder': folders, 'message': error})
+
+
 def admin_page(request):
-    profile = UserProfile.objects.filter(user = request.user)[0]
-    if not profile.is_admin :
+    if check_user_fail(request):
         return render(request, 'admin/reject.html')
 
     Groups = UserGroup.objects.all()[:20]
@@ -61,16 +107,25 @@ def admin_page(request):
     return render(request, 'admin/main.html', context)
 
 def admin_user(request, user_id):
+    if check_user_fail(request):
+        return render(request, 'admin/reject.html')
+
     context = {'u' : UserProfile.objects.filter(pk=user_id)[0]}
     return render(request, 'admin/user.html', context)
 
 def admin_group(request, group_id):
+    if check_user_fail(request):
+        return render(request, 'admin/reject.html')
+
     return render(request, 'admin/group.html')
 
 @sensitive_post_parameters()
 @csrf_protect
 @never_cache
 def admin_creategroup(request, creation_form=GroupCreationForm):
+    if check_user_fail(request):
+        return render(request, 'admin/reject.html')
+
     if request.method == "POST":
         form = creation_form(data=request.POST)
         if form.is_valid() :
@@ -85,8 +140,42 @@ def admin_creategroup(request, creation_form=GroupCreationForm):
     return render(request, "admin/creategroup.html", context)
 
 def admin_deleteuser(request, user_id):
+    if check_user_fail(request):
+        return render(request, 'admin/reject.html')
+
     u = UserProfile.objects.filter(pk=user_id)[0]
     u_u = u.user
     u.delete()
     u_u.delete()
     return render(request, 'admin/action_complete.html')
+
+def admin_makeadmin(request, user_id):
+    if check_user_fail(request):
+        return render(request, 'admin/reject.html')
+
+    u = UserProfile.objects.filter(pk=user_id)[0]
+    u.administrator = 1
+    u.save()
+    return render(request, 'admin/action_complete.html')
+
+def check_user_fail(request):
+    try:
+        profile = UserProfile.objects.filter(user = request.user)[0]
+    except IndexError:
+        return True
+    if not profile.is_admin :
+        return True
+
+@login_required(login_url="/accounts/login/")
+def add_group(request, creation_form=UserGroupCreationForm):
+    if request.method == "POST":
+        form = creation_form(data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect('/')
+    else:
+        form=creation_form(request)
+    context = {
+        'form' : form,
+    }
+    return render(request, "add_group.html")
